zm: ♻️ Move `PropertyEmitsChangedSignal` to utils.rs
